+++ yash-syntax/src/parser/lex/core.rs
@@ -212,7 +212,10 @@ impl<'a> LexerCore<'a> {
 
             // Read more input
             let index = self.next_index();
-            match self.input.next_line(&Context).await {
+            let context = Context {
+                is_first_line: self.raw_code.value.borrow().is_empty(),
+            };
+            match self.input.next_line(&context).await {
                 Ok(line) => {
                     if line.is_empty() {
                         // End of input
@@ -879,6 +882,34 @@ mod tests {
         assert_eq!(e.location.code.start_line_number, line);
         assert_eq!(e.location.code.source, Source::Unknown);
         assert_eq!(e.location.range, 0..0);
+    }
+
+    #[test]
+    fn lexer_core_peek_char_context_is_first_line() {
+        // In this test case, this mock input function will be called twice.
+        struct InputMock {
+            first: bool,
+        }
+        #[async_trait::async_trait(?Send)]
+        impl Input for InputMock {
+            async fn next_line(&mut self, context: &Context) -> crate::input::Result {
+                assert_eq!(context.is_first_line(), self.first);
+                self.first = false;
+                Ok("\n".to_owned())
+            }
+        }
+
+        let input = InputMock { first: true };
+        let line = NonZeroU64::new(42).unwrap();
+        let mut lexer = LexerCore::new(Box::new(input), line, Source::Unknown);
+
+        let peek = lexer.peek_char().now_or_never().unwrap();
+        assert_matches!(peek, Ok(PeekChar::Char(_)));
+        lexer.consume_char();
+
+        let peek = lexer.peek_char().now_or_never().unwrap();
+        assert_matches!(peek, Ok(PeekChar::Char(_)));
+        lexer.consume_char();
     }
 
     #[test]